Update README.md
Generating quantum music using Qiskit. We have created a platform designed to help the general public understand the quantum phenomena of superposition and quantum computation, all using the language of music instead of mathematics. Each musical note is assigned a qubit. Using an interactive quantum piano interface (built using Pygame), the user can press a key on their keyboard, which plays a piano key on the screen. This generates the relevant quantum logic gate on the screen. Changing the volume will demonstrate superposition, as maximum volume = |1> and minimum volume (i.e. no sound) = |0>, so any volume in between will be a superposition of the two, which is not possible classically. This interactive platform is designed to teach the user about quantum circuitry, so that when the process is reversed, they can appreciate what they are listening to. Therefore, the finally is an audio demo in which a quantum algorithm is translated into music, so you can listen to what a quantum algorithm sounds like.